panorama de la página de inicio mejorada
