+++ src/components/footer/footer.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react"
 import { Link } from "gatsby"
 import "./footer.css"
-import Logo from "../../../static/logo.svg"
+import Logo from "../../../static/logo.png"
 import "../../../node_modules/@fortawesome/fontawesome-free/css/all.css"
 
 const Footer = () => {
