+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Link } from "gatsby"
 import Navbar from "../components/navbar/navbar"
 import CalendarA from "../components/calendar/CalendarA"
-import Header from "../components/header/header"
+import PanoramaInicio from "../components/PanoramaInicio/panorama"
 import Testimonios from "../components/testimonios/testimonios"
 import Footer from "../components/footer/footer"
 
@@ -10,7 +10,7 @@ export default function Home() {
   return (
     <div>
         <Navbar />
-        <Header />
+        <PanoramaInicio />
         <h1> Academia Temple </h1> 
         <p> Mejorando la educación a nivel global </p> 
         <CalendarA></CalendarA>
